fix(parsing): extraire le texte des documents avec un parseur, pas des regex
CodeQL remonte trois alertes "incomplete multi-character sanitization" sur
extractBlockText : une passe de regex qui retire "<script>" recompose une
balise vivante à partir de "<scr<script>ipt>", même chose pour les
commentaires. Le HTML vient d'une source externe, la garantie ne tient pas.

extractBlockText passe donc par cheerio, déjà utilisé ailleurs dans le
projet, et part dans son propre module : html-utils reste sans dépendance
pour les modules qui n'ont besoin que des helpers légers.

Effet de bord assumé : le parseur décode les entités fidèlement, donc
&rsquo; devient U+2019 au lieu de l'apostrophe droite que la table maison
substituait. C'est le caractère que l'AN utilise réellement.
